Fixed wrong names in Visual Studio projects; added helper that creates a temporary file and deletes it as soon as the instance goes out of scope (very useful for writing unit tests - resisted writing this at first, but since I now have minor file system stuff in Nuclex.Support anyway...)
git-svn-id: file:///srv/devel/repo-conversion/nusu@332 d2e56fa2-650e-0410-a79f-9358c0239efd

+++ Source/Platform/WindowsApi.cpp
@@ -139,7 +139,7 @@ namespace Nuclex { namespace Support { namespace Platform {
       );
       if(errorMessageLength == 0) {
         std::string message(u8"Windows API error ");
-        Text::lexical_append(message, errorCode);
+        Text::lexical_append(message, static_cast<std::uint32_t>(errorCode));
         return message;
       }
     }
@@ -169,7 +169,7 @@ namespace Nuclex { namespace Support { namespace Platform {
     // If the error message is empty, return a generic one
     if(size == 0) {
       std::string message(u8"Windows API error ");
-      Text::lexical_append(message, errorCode);
+      Text::lexical_append(message, static_cast<std::uint32_t>(errorCode));
       return message;
     } else { // Error message had content, return it
       return std::string(&utf8ErrorMessage[0], size);
